+++ MacPacker/Core/Constants.swift
@@ -9,4 +9,7 @@ class Constants {
     static let supportMail: String = "apps@sarensw.com"
     static let otherAppFileFillet: String = "FileFillet"
     static let otherAppGitHub: String = "GitHub"
+    
+    static let appStoreReviewLink: String = "https://apps.apple.com/app/id6473273874?action=write-review"
+    static let gitHubLink: String = "https://github.com/sarensw/MacPacker"
 }